Hide fake file descriptor from dataset details
